dataValidation rows/cols *must* be contiguous.
Data validation rows and columns are specified via a `startRow` and
`endRow` arguments (R/wrappers.R:3015), so the `rows` and `cols`
parameters to `dataValidation()` must be contiguous. This updates the
documentation. `dataValidation()` could be updated to take
non-contiguous rows/cols, but based on the current method, would need
to add an entry for *each* row/column desired.