Fix Bug #47: Protected workbooks generate Ecxel-error when loaded with openxlsx => Fix order of workbook subelements
The workbook protection settings (part of the workbook XML file) were incorrectly implemented, so that loading a protected workbook in openxlsx caused the workbookProtection XML tag to be appended to the workbook tag. The spec, however, demands a given order of the elements, so Excel printed a warning when a protected file was read and re-written using openxlsx.

The fix is to add a NULL workbookProtection to the default fields of the workbook object, so it will always be in the correct position. Reading then simply needs to read the corresponding tag from the XML, and writing does not need anything special.

Fixes bug #47