Add detail to the error when failing to import dates
Fixes a bug where importing a sheet with a date colum that also
contained invalid dates would fail with a c++ exception and no other
info.
Adds detail to the error display when failing. Shows the user the text
that failed to import correctly along with the row and column numbers.
Rethrows the exception so that the process still fails to avoid
unanticipated behavior if someone doesn't look at the error messages.
Fixes #172