apparently comments require vmlDrawings, but when the comment is created, there is no object created in vml/vml_rels. Henc our testing on vml/vml_rels always fails.
As a fix, this commit re-imports the vmlDrawing written in writeDrawingVML. This might overwrite whatever was in vml prior to the import, though I'm not sure that it was used previously. If I'm not mistaken, the comment was written and the vml could have only been appended to the file.

This requires testing and maybe another solution saving imported vml content.