Bug #330: writeData with a date creates a new number format style, which potentially needs a unique number format ID
Make sure that writeData with a date creates a style that has no collisions of its numFmtId with existing styles. By default, a style with numFmdId=165 (hard-coded!) is created by createStyle. When inserting that style into a workbook, we need to assign it a new, unique numFmtId if the workbook already has some custom number formats.
The explicitly-called addStyle function checked for this case, but the implicitly-called classStyles function did not.

Fixes #330.