replace file.exists with dir.exists
My suggestion is to use the  dir.exists function to check if a file object exists and if it is a directory. See ?dir.exists for base::dir.exists -
"dir.exists checks that the paths exist (in the same sense as file.exists) and are directories."

In the documentation for file.exists, there's system-dependend limitations mentioned, in particular "What constitutes a ‘file’ is system-dependent, but should include directories. (However, directory names must not include a trailing backslash or slash on Windows.)":
https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/files